Fix mismatch between Client and Server queue names
The client is removing the MQ prefix from the QueueName before creating the connection.
The server doesn't do this, therefore they are listening and sending to different queues.

If no prefix is wanted this can be configured by setting QueueNames.MqPrefix to an empty string.
Therefore I see no reason to always remove the MqPrefix.